Use image types from `imgref` instead of `image`.
`image` publishes incompatible versions not infrequently.
Therefore, it is not a good choice as a dependency that appears in our
public API. `imgref`, on the other hand, is explicitly minimal, and has
existed for years at 1.x.

Unfortunately, this means that users of `rendiff` have to perform
conversions, but they would have had to do that anyway if their version
of `image` is not the same as ours.

Implements <https://github.com/kpreid/rendiff/issues/3>.