Optimize: In `half_diff`, use `extend()` per image row instead of `push()` per pixel.
I intended this to be preliminary refactoring, but it turns out to make
a large performance improvement (110 ms to 75 ms) by itself. The
improvement is fragile and goes away with other small changes, which
suggests that it might be an accident of machine code generation more
than a reliable improvement, but getting rid of the `push()`es seems
like a better state for future work, so even if they're really
just equivalent, I'll take this.